Bug: update the `App.jsx` to fix the issue of the viewport
This commit has ONE CHANGE AND ONLY ONE:
Fixed the issue of the viewport by adding the `useExpandToFooter` hook
to the `App.jsx` file. This hook is responsible for expanding the
viewport to accommodate the footer. This code was written by
`@Hamada-Gado`. Thanks for the contribution.

This commit will be commited onto a local branch called `feature/admin`,
but then will be merged right afterwards into the `main` branch.

It is advised that all contributors update their local branches
to the latest `main` branch before pushing their changes.

!!! READ THIS !!!