Feat: add delete functionality
Its not working yet, as the issue is that we need to use `useEffect` to
do a post request to the server to update the db file. the issue is that
we cannot use any hooks inside a function, so we need to refactor the
`useDelete` hook to be a function that returns a function that we can
use in the `onClick` event.

As a side note i am using dependency injection to pass the `setItems`
function to the `useDelete` hook, this is a good practice as it makes
the hook more reusable.